added graph test mode

+++ ddh/utils_graph.py
@@ -1,11 +1,12 @@
 from functools import lru_cache
 from glob import glob
 import os
+
+from settings.ctx import g_graph_test_mode
 from utils.logs import lg_gra as lg
 import pandas as pd
 import dateutil.parser as dp
 from os.path import basename
-
 from utils.ddh_shared import ddh_get_absolute_application_path
 
 
@@ -24,6 +25,14 @@ def graph_get_fol_list() -> list:
     get list of absolute paths of "dl_files/<mac>" folders
     """
     d = ddh_get_absolute_application_path() + '/dl_files'
+    if g_graph_test_mode:
+        fol_ls = [
+            d + '/11-22-33-44-55-66',
+            d + '/00-00-00-00-00-00',
+        ]
+        return fol_ls
+
+    d = ddh_get_absolute_application_path() + '/dl_files'
     if os.path.isdir(d):
         f_l = [f.path for f in os.scandir(d) if f.is_dir()]
         # remove 'ddh_vessel' folders
@@ -33,7 +42,10 @@ def graph_get_fol_list() -> list:
 
 # read GRAPH_REQ_JSON_FILE, it has the FULL ABSOLUTE folder path to plot
 def graph_get_abs_fol_req_file():
-    # file written by DDS when requesting a graph
+    if g_graph_test_mode:
+        return
+
+    # file written by DDS_BLE when requesting a graph
     with open(GRAPH_REQ_JSON_FILE) as f:
         fol = f.read().strip()
     if not os.path.exists(fol):
@@ -44,10 +56,14 @@ def graph_get_abs_fol_req_file():
 
 
 def graph_check_fol_req_file():
+    if g_graph_test_mode:
+        return
     return os.path.exists(GRAPH_REQ_JSON_FILE)
 
 
 def graph_delete_fol_req_file():
+    if g_graph_test_mode:
+        return
     try:
         os.unlink(GRAPH_REQ_JSON_FILE)
     except (Exception, ) as ex:
@@ -55,6 +71,8 @@ def graph_delete_fol_req_file():
 
 
 def graph_set_fol_req_file(mac):
+    if g_graph_test_mode:
+        return
     try:
         with open(GRAPH_REQ_JSON_FILE, "w") as f:
             fol = mac.replace(':', '-')
